fixed issue 64 AfterCapture tasks behavior is modified as follows:
if "Copy image to clipboard" and "Save to file" after capture tasks are enabled then the file is saved and the image is copied to the clipboard. 

if "Save to file" after capture task is selected, the file path will be copied to the clipboard. 

if "Upload to remove host" after capture task is selected, then URL will be copied to the clipboard.